feat: CLIN-1529 Reorganization Cypress tests
