+++ cypress/e2e/Facettes/PagePrescriptions.cy.ts
@@ -7,138 +7,124 @@ beforeEach(() => {
   cy.login(Cypress.env('username_DG_CHUSJ_CUSM_CHUS'), Cypress.env('password'));
 });
 
-describe('Page des prescriptions et des requêtes', () => {
-  describe('Filtrer avec les facettes', () => {
+describe('Page des prescriptions et requêtes - Filtrer avec les facettes', () => {
 
-    beforeEach(() => {
-      cy.visitPrescriptionsPage();
-    });
+  beforeEach(() => {
+    cy.visitPrescriptionsPage();
+  });
 
-    describe('Statut', () => {
-      it('Approuvée', () => {
-        cy.checkValueFacet(0, 'Approuvée');
+  it('Statut - Approuvée', () => {
+    cy.checkValueFacet(0, 'Approuvée');
 
-        cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+    cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
 
-        cy.get('body').contains('Prescriptions (1)').should('exist');
-        cy.get('body').contains('Requêtes (3)').should('exist');
-      });
-    });
-      
-    describe('Statut des requêtes', () => {
-      it('Complétée', () => {
-        cy.checkValueFacet(1, 'Complétée');
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains('Requêtes (3)').should('exist');
+  });
+  
+  it('Statut des requêtes - Complétée', () => {
+    cy.checkValueFacet(1, 'Complétée');
 
-        cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+    cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
 
-        cy.get('body').contains('Prescriptions (1)').should('exist');
-        cy.get('body').contains('Requêtes (3)').should('exist');
-      });
-    });
-      
-    describe('Analyse', () => {
-      it('Malignant Hyperthermia (HYPM)', () => {
-        cy.checkValueFacet(2, 'HYPM');
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains('Requêtes (3)').should('exist');
+  });
+  
+  it('Analyse - Malignant Hyperthermia (HYPM)', () => {
+    cy.checkValueFacet(2, 'HYPM');
 
-        cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+    cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
 
-        cy.get('body').contains('Prescriptions 0').should('exist');
-        cy.get('body').contains('Requêtes 0').should('exist');
-      });
+    cy.get('body').contains('Prescriptions 0').should('exist');
+    cy.get('body').contains('Requêtes 0').should('exist');
+  });
 
-      it('Congenital Myopathies (MYOC)', () => {
-        cy.checkValueFacet(2, 'MYOC');
+  it('Analyse - Congenital Myopathies (MYOC)', () => {
+    cy.checkValueFacet(2, 'MYOC');
 
-        cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+    cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
 
-        cy.get('body').contains('Prescriptions 0').should('exist');
-        cy.get('body').contains('Requêtes 0').should('exist');
-      });
+    cy.get('body').contains('Prescriptions 0').should('exist');
+    cy.get('body').contains('Requêtes 0').should('exist');
+  });
 
-      it('Global Developmental Delay / Intellectual Deficiency (Trio) (RGDI)', () => {
-        cy.checkValueFacet(2, 'RGDI');
+  it('Analyse - Global Developmental Delay / Intellectual Deficiency (Trio) (RGDI)', () => {
+    cy.checkValueFacet(2, 'RGDI');
 
-        cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+    cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
 
-        cy.get('body').contains('Prescriptions (1)').should('exist');
-        cy.get('body').contains('Requêtes (3)').should('exist');
-      });
-    });
-      
-    describe('Laboratoire (LDM)', () => {
-      it('LDM-CHUSJ', () => {
-        cy.checkValueFacet(3, 'LDM-CHUSJ');
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains('Requêtes (3)').should('exist');
+  });
+  
+  it('Laboratoire (LDM) - LDM-CHUSJ', () => {
+    cy.checkValueFacet(3, 'LDM-CHUSJ');
 
-        cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+    cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
 
-        cy.get('body').contains('Prescriptions (1)').should('exist');
-        cy.get('body').contains('Requêtes (3)').should('exist');
-      });
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains('Requêtes (3)').should('exist');
+  });
 
-      it('LDM-CHUS', () => {
-        cy.checkValueFacet(3, /^LDM-CHUS$/);
+  it('Laboratoire (LDM) - LDM-CHUS', () => {
+    cy.checkValueFacet(3, /^LDM-CHUS$/);
 
-        cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+    cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
 
-        cy.get('body').contains('Prescriptions 0').should('exist');
-        cy.get('body').contains('Requêtes 0').should('exist');
-      });
+    cy.get('body').contains('Prescriptions 0').should('exist');
+    cy.get('body').contains('Requêtes 0').should('exist');
+  });
 
-      it('LDM-CUSM', () => {
-        cy.checkValueFacet(3, 'LDM-CUSM');
+  it('Laboratoire (LDM) - LDM-CUSM', () => {
+    cy.checkValueFacet(3, 'LDM-CUSM');
 
-        cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+    cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
 
-        cy.get('body').contains('Prescriptions 0').should('exist');
-        cy.get('body').contains('Requêtes 0').should('exist');
-      });
-    });
-      
-    describe('Établissement prescripteur', () => {
-      it('CHUSJ', () => {
-        cy.checkValueFacet(4, 'CHUSJ');
+    cy.get('body').contains('Prescriptions 0').should('exist');
+    cy.get('body').contains('Requêtes 0').should('exist');
+  });
+  
+  it('Établissement prescripteur - CHUSJ', () => {
+    cy.checkValueFacet(4, 'CHUSJ');
 
-        cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+    cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
 
-        cy.get('body').contains('Prescriptions (1)').should('exist');
-        cy.get('body').contains('Requêtes (3)').should('exist');
-      });
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains('Requêtes (3)').should('exist');
+  });
 
-      it('CHUS', () => {
-        cy.checkValueFacet(4, /^CHUS$/);
+  it('Établissement prescripteur - CHUS', () => {
+    cy.checkValueFacet(4, /^CHUS$/);
 
-        cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+    cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
 
-        cy.get('body').contains('Prescriptions 0').should('exist');
-        cy.get('body').contains('Requêtes 0').should('exist');
-      });
+    cy.get('body').contains('Prescriptions 0').should('exist');
+    cy.get('body').contains('Requêtes 0').should('exist');
+  });
 
-      it('CUSM', () => {
-        cy.checkValueFacet(4, 'CUSM');
+  it('Établissement prescripteur - CUSM', () => {
+    cy.checkValueFacet(4, 'CUSM');
 
-        cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+    cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
 
-        cy.get('body').contains('Prescriptions 0').should('exist');
-        cy.get('body').contains('Requêtes 0').should('exist');
-      });
-    });
-      
-    describe('Prénatal', () => {
-      it('False', () => {
-        cy.intercept('POST', '**/graphql').as('getPOSTgraphql');
+    cy.get('body').contains('Prescriptions 0').should('exist');
+    cy.get('body').contains('Requêtes 0').should('exist');
+  });
+  
+  it('Prénatal - False', () => {
+    cy.intercept('POST', '**/graphql').as('getPOSTgraphql');
 
-        cy.get('div[class="Filter_facetCollapse__ft2Q2"]').eq(5)
-          .find('div[class*="ToggleFilter_radio"]').contains('False')
-          .find('[type="radio"]').check({force: true});
+    cy.get('div[class="Filter_facetCollapse__ft2Q2"]').eq(5)
+      .find('div[class*="ToggleFilter_radio"]').contains('False')
+      .find('[type="radio"]').check({force: true});
 
-        cy.wait('@getPOSTgraphql', {timeout: 5000});
-        cy.wait('@getPOSTgraphql', {timeout: 5000});
+    cy.wait('@getPOSTgraphql', {timeout: 5000});
+    cy.wait('@getPOSTgraphql', {timeout: 5000});
 
-        cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+    cy.get('input[class="ant-input"]').first().type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
 
-        cy.get('body').contains('Prescriptions (1)').should('exist');
-        cy.get('body').contains('Requêtes (3)').should('exist');
-      });
-    });
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains('Requêtes (3)').should('exist');
   });
 });
