+++ cypress/e2e/Recherches/PagePrescriptions.cy.ts
@@ -7,205 +7,203 @@ beforeEach(() => {
   cy.login(Cypress.env('username_DG_CHUSJ_CUSM_CHUS'), Cypress.env('password'));
 });
 
-describe('Page des prescriptions et des requêtes', () => {
-  describe('Rechercher des prescriptions', () => {
+describe('Page des prescriptions et requêtes - Rechercher des prescriptions', () => {
 
-    beforeEach(() => {
-      cy.visitPrescriptionsPage();
-    });
-
-    it('Par numéro de prescription', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
-
-      cy.get('body').contains('Prescriptions (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
-    });
-
-    it('Par numéro de requête du cas-index', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.requestProbId, {force: true});
-
-      cy.get('body').contains('Prescriptions (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
-    });
-
-    it('Par numéro de dossier du cas-index', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.mrnProb, {force: true});
-
-      cy.get('body').contains('Prescriptions (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
-    });
-
-    it('Par numéro de patient du cas-index', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.patientProbId, {force: true});
-
-      cy.get('body').contains('Prescriptions (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
-    });
-
-    it('Par numéro d\'échantillon du cas-index', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.sampleProbId, {force: true});
-
-      cy.get('body').contains('Prescriptions (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
-    });
-
-    it('Par numéro de requête de la mère', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.requestMthId, {force: true});
-
-      cy.get('body').contains('Prescriptions (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
-    });
-
-    it('Par numéro de dossier de la mère', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.mrnMth, {force: true});
-
-      cy.get('body').contains('Prescriptions (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
-    });
-
-    it('Par numéro de patient de la mère', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.patientMthId, {force: true});
-
-      cy.get('body').contains('Prescriptions (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
-    });
-
-    it('Par numéro d\'échantillon de la mère', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.sampleMthId, {force: true});
-
-      cy.get('body').contains('Prescriptions (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
-    });
-
-    it('Par numéro de requête du père', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.requestFthId, {force: true});
-
-      cy.get('body').contains('Prescriptions (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
-    });
-
-    it('Par numéro de dossier du père', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.mrnFth, {force: true});
-
-      cy.get('body').contains('Prescriptions (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
-    });
-
-    it('Par numéro de patient du père', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.patientFthId, {force: true});
-
-      cy.get('body').contains('Prescriptions (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
-    });
-
-    it('Par numéro d\'échantillon du père', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.sampleFthId, {force: true});
-
-      cy.get('body').contains('Prescriptions (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
-    });
+  beforeEach(() => {
+    cy.visitPrescriptionsPage();
   });
 
-  describe('Rechercher des requêtes', () => {
+  it('Par numéro de prescription', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
 
-    beforeEach(() => {
-      cy.visitPrescriptionsPage();
-      cy.get('div[id*="tab-requests"]').click({force: true});
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
+  });
 
-      cy.resetColumns(1);
-    });
+  it('Par numéro de requête du cas-index', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.requestProbId, {force: true});
 
-    it('Par numéro de prescription', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
+  });
 
-      cy.get('body').contains('Requêtes (3)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestProbId).should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestMthId).should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestFthId).should('exist');
-    });
+  it('Par numéro de dossier du cas-index', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.mrnProb, {force: true});
 
-    it('Par numéro de requête du cas-index', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.requestProbId, {force: true});
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
+  });
 
-      cy.get('body').contains('Requêtes (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestProbId).should('exist');
-    });
+  it('Par numéro de patient du cas-index', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.patientProbId, {force: true});
 
-    it('Par numéro de dossier du cas-index', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.mrnProb, {force: true});
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
+  });
 
-      cy.get('body').contains('Requêtes (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestProbId).should('exist');
-    });
+  it('Par numéro d\'échantillon du cas-index', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.sampleProbId, {force: true});
 
-    it('Par numéro de patient du cas-index', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.patientProbId, {force: true});
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
+  });
 
-      cy.get('body').contains('Requêtes (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestProbId).should('exist');
-    });
+  it('Par numéro de requête de la mère', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.requestMthId, {force: true});
 
-    it('Par numéro d\'échantillon du cas-index', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.sampleProbId, {force: true});
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
+  });
 
-      cy.get('body').contains('Requêtes (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestProbId).should('exist');
-    });
+  it('Par numéro de dossier de la mère', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.mrnMth, {force: true});
 
-    it('Par numéro de requête de la mère', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.requestMthId, {force: true});
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
+  });
 
-      cy.get('body').contains('Requêtes (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestMthId).should('exist');
-    });
+  it('Par numéro de patient de la mère', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.patientMthId, {force: true});
 
-    it('Par numéro de dossier de la mère', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.mrnMth, {force: true});
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
+  });
 
-      cy.get('body').contains('Requêtes (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestMthId).should('exist');
-    });
+  it('Par numéro d\'échantillon de la mère', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.sampleMthId, {force: true});
 
-    it('Par numéro de patient de la mère', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.patientMthId, {force: true});
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
+  });
 
-      cy.get('body').contains('Requêtes (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestMthId).should('exist');
-    });
+  it('Par numéro de requête du père', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.requestFthId, {force: true});
 
-    it('Par numéro d\'échantillon de la mère', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.sampleMthId, {force: true});
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
+  });
 
-      cy.get('body').contains('Requêtes (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestMthId).should('exist');
-    });
+  it('Par numéro de dossier du père', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.mrnFth, {force: true});
 
-    it('Par numéro de requête du père', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.requestFthId, {force: true});
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
+  });
 
-      cy.get('body').contains('Requêtes (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestFthId).should('exist');
-    });
+  it('Par numéro de patient du père', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.patientFthId, {force: true});
 
-    it('Par numéro de dossier du père', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.mrnFth, {force: true});
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
+  });
 
-      cy.get('body').contains('Requêtes (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestFthId).should('exist');
-    });
+  it('Par numéro d\'échantillon du père', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.sampleFthId, {force: true});
 
-    it('Par numéro de patient du père', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.patientFthId, {force: true});
+    cy.get('body').contains('Prescriptions (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.prescriptionId).should('exist');
+  });
+});
 
-      cy.get('body').contains('Requêtes (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestFthId).should('exist');
-    });
+describe('Page des prescriptions et requêtes - Rechercher des requêtes', () => {
 
-    it('Par numéro d\'échantillon du père', () => {
-      cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.sampleFthId, {force: true});
+  beforeEach(() => {
+    cy.visitPrescriptionsPage();
+    cy.get('div[id*="tab-requests"]').click({force: true});
 
-      cy.get('body').contains('Requêtes (1)').should('exist');
-      cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestFthId).should('exist');
-    });
+    cy.resetColumns(1);
+  });
+
+  it('Par numéro de prescription', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.prescriptionId, {force: true});
+
+    cy.get('body').contains('Requêtes (3)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestProbId).should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestMthId).should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestFthId).should('exist');
+  });
+
+  it('Par numéro de requête du cas-index', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.requestProbId, {force: true});
+
+    cy.get('body').contains('Requêtes (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestProbId).should('exist');
+  });
+
+  it('Par numéro de dossier du cas-index', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.mrnProb, {force: true});
+
+    cy.get('body').contains('Requêtes (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestProbId).should('exist');
+  });
+
+  it('Par numéro de patient du cas-index', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.patientProbId, {force: true});
+
+    cy.get('body').contains('Requêtes (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestProbId).should('exist');
+  });
+
+  it('Par numéro d\'échantillon du cas-index', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.sampleProbId, {force: true});
+
+    cy.get('body').contains('Requêtes (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestProbId).should('exist');
+  });
+
+  it('Par numéro de requête de la mère', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.requestMthId, {force: true});
+
+    cy.get('body').contains('Requêtes (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestMthId).should('exist');
+  });
+
+  it('Par numéro de dossier de la mère', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.mrnMth, {force: true});
+
+    cy.get('body').contains('Requêtes (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestMthId).should('exist');
+  });
+
+  it('Par numéro de patient de la mère', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.patientMthId, {force: true});
+
+    cy.get('body').contains('Requêtes (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestMthId).should('exist');
+  });
+
+  it('Par numéro d\'échantillon de la mère', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.sampleMthId, {force: true});
+
+    cy.get('body').contains('Requêtes (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestMthId).should('exist');
+  });
+
+  it('Par numéro de requête du père', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.requestFthId, {force: true});
+
+    cy.get('body').contains('Requêtes (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestFthId).should('exist');
+  });
+
+  it('Par numéro de dossier du père', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.mrnFth, {force: true});
+
+    cy.get('body').contains('Requêtes (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestFthId).should('exist');
+  });
+
+  it('Par numéro de patient du père', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.patientFthId, {force: true});
+
+    cy.get('body').contains('Requêtes (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestFthId).should('exist');
+  });
+
+  it('Par numéro d\'échantillon du père', () => {
+    cy.get('[data-cy="PrescriptionsSearch"]').type(epCHUSJ_ldmCHUSJ.sampleFthId, {force: true});
+
+    cy.get('body').contains('Requêtes (1)').should('exist');
+    cy.get('body').contains(epCHUSJ_ldmCHUSJ.requestFthId).should('exist');
   });
 });
