+++ cypress/support/index.d.ts
@@ -10,7 +10,7 @@ declare namespace Cypress {
     logout(): cy & CyEventEmitter;
     resetColumns(eq: number): cy & CyEventEmitter;
     typeAndIntercept(selector: string, text: string, methodHTTP: string, routeMatcher: string, nbCalls: number): cy & CyEventEmitter;
-    validateDictionnary(facetTitle: string, facetRank: number, dictionnary: string[]): cy & CyEventEmitter;
+    validateDictionnary(section: string, facetTitle: string, facetRank: number, dictionnary: string[]): cy & CyEventEmitter;
     visitAndIntercept(url: string, methodHTTP: string, routeMatcher: string, nbCalls: number): cy & CyEventEmitter;
     visitArchivesPatientPage(patientId: string): cy & CyEventEmitter;
     visitBioinformaticsAnalysisPage(bioAnalysisId: string): cy & CyEventEmitter;
